requirements.txt
APScheduler==3.10.1
autopep8==2.0.0
beautifulsoup4==4.12.2
blinker==1.5
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.24.3
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.1
sentry-sdk==1.19.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.45
threadpoolctl==3.1.0
tomli==2.0.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.2.2
zipp==3.11.0
